开始重构代码，未完成

+++ ZeroMS-Server/OServerCore_MsgSlot.cpp
@@ -214,25 +214,39 @@ void OServerCore::AskUserList(OClient::Connect *connect,QString listname,QString
     protocol.UserList(connect,listname,operation,allList);
 }
 
-void OServerCore::ModifyUserList(OClient::Connect *connect,QString uname,bool isAddOrRemove)
+void OServerCore::ModifyUserList(OClient::Connect *connect,QString listname,QString uname, bool isAddOrRemove, QString message)
 {
     if(connect->client->isLoged)
     {//如果已经登录
-        bool isGroup=(uname.left(1)=="*")?true:false;
-        if(isAddOrRemove && !isGroup)
-        {//如果是添加用户
-            if(db.checkUser(uname))
-            {//如果这个用户存在
-                db.ModifyUserList(connect->client->uname,uname,true);
+        if(listname==connect->client->uname)
+        {
+            bool isGroup=(uname.left(1)=="*")?true:false;
+            if(isAddOrRemove && !isGroup)
+            {//如果是添加用户
+                if(db.checkUser(uname))
+                {//如果这个用户存在
+                    db.ModifyUserList(connect->client->uname,uname,true);
+                }
+            }
+            else if(!isGroup)
+            {//如果是删除用户
+                db.ModifyUserList(connect->client->uname,uname,false);
+            }
+            else
+            {//如果是删除群
+                db.removeGroupMember(uname,connect->client->uname);
             }
         }
-        else if(!isGroup)
-        {//如果是删除用户
-            db.ModifyUserList(connect->client->uname,uname,false);
-        }
         else
-        {//如果是删除群
-            db.removeGroupMember(uname,connect->client->uname);
+        {
+            if(db.checkGroup(listname))
+            {
+
+            }
+            else
+            {
+                protocol.Unknown(connect);
+            }
         }
     }
     else
